Move IEventEmitter to OmniSharp.Eventing namespace and move IEventEmitter implementations into OmniSharp.Host
