+++ src/OmniSharp.DotNet/Cache/ProjectStatesCache.cs
@@ -6,9 +6,9 @@ using Microsoft.DotNet.ProjectModel;
 using Microsoft.Extensions.Logging;
 using NuGet.Frameworks;
 using OmniSharp.DotNet.Models;
+using OmniSharp.Eventing;
 using OmniSharp.Models.Events;
 using OmniSharp.Models.ProjectInformation;
-using OmniSharp.Services;
 
 namespace OmniSharp.DotNet.Cache
 {
