+++ src/OmniSharp.DotNetTest/Services/BaseTestService.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
+using OmniSharp.Eventing;
 using OmniSharp.Services;
 
 namespace OmniSharp.DotNetTest.Services
