+++ src/OmniSharp.DotNetTest/Services/BaseTestService`2.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
+using OmniSharp.Eventing;
 using OmniSharp.Mef;
 using OmniSharp.Models;
 using OmniSharp.Services;
