+++ src/OmniSharp.DotNetTest/Services/DebugTestService.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
 using OmniSharp.DotNetTest.Models;
+using OmniSharp.Eventing;
 using OmniSharp.Mef;
 using OmniSharp.Services;
