+++ src/OmniSharp.DotNetTest/Services/RunTestService.cs
@@ -2,6 +2,7 @@ using System.Composition;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
 using OmniSharp.DotNetTest.Models;
+using OmniSharp.Eventing;
 using OmniSharp.Mef;
 using OmniSharp.Services;
