+++ src/OmniSharp.Roslyn/ProjectEventForwarder.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Composition;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
+using OmniSharp.Eventing;
 using OmniSharp.Models.Events;
 using OmniSharp.Models.ProjectInformation;
 using OmniSharp.Services;
