+++ tests/OmniSharp.Roslyn.CSharp.Tests/DiagnosticsV2Facts.DiagnosticTestEmitter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using OmniSharp.Eventing;
 using OmniSharp.Models.Diagnostics;
-using OmniSharp.Services;
 
 namespace OmniSharp.Roslyn.CSharp.Tests
 {
